test: split shared explorer e2e test in twain
